Updated dependencies, more refactoring
