+++ client/package.json
@@ -4,7 +4,7 @@
 	"description": "Detect security mistakes as you type, and get real time details and fix suggestions",
 	"author": "Microsoft Corporation",
 	"license": "MIT",
-	"version": "0.2.1",
+	"version": "0.5.0",
 	"publisher": "MS-DevSkim",
 	"repository": {
 		"type": "git",
