+++ server/package.json
@@ -1,7 +1,7 @@
 {
   "name": "devskim-server",
   "description": "Devskim Language Server component",
-  "version": "0.2.1",
+  "version": "0.5.0",
   "author": "Security Engineering",
   "publisher": "MS-DevSkim",
   "private": true,
@@ -14,13 +14,13 @@
   },
   "dependencies": {
     "glob-promise": "^3.4.0",
-    "lodash": "^4.17.11",
+    "lodash": "^4.17.15",
     "mkdirp": "^0.5.1",
     "node-dir": ">=0.1.17",
     "vscode-languageserver": "^5.2.1",
     "vscode-languageserver-protocol": "^3.14.1",
-    "vscode-uri": "^1.0.6",
-    "xregexp": "^3.2.0"
+    "vscode-uri": "^2.0.3",
+    "xregexp": "^4.2.4"
   },
   "devDependencies": {
     "@types/jest": "^24.0.11",
